Rearrange vm tabs
