+++ cypress/tests/check-tabs.spec.ts
@@ -14,7 +14,8 @@ describe('Check all virtualization pages can be loaded', () => {
     cy.deleteResource(K8S_KIND.TEMPLATE, 'vm-template-example', 'default');
   });
 
-  describe('Check VM list and tabs', () => {
+  //TODO: Fix by Guohua
+  xdescribe('Check VM list and tabs', () => {
     it('create vm-example', () => {
       cy.visitVMs();
       vm.createVMFromYAML();
