+++ locales/en/plugin__kubevirt-plugin.json
@@ -351,7 +351,7 @@
   "Developer preview": "Developer preview",
   "Developer preview features are not intended to be used in production environments. The clusters deployed with the developer preview features are considered to be development clusters and are not supported through the Red Hat Customer Portal case management system.": "Developer preview features are not intended to be used in production environments. The clusters deployed with the developer preview features are considered to be development clusters and are not supported through the Red Hat Customer Portal case management system.",
   "Device name": "Device name",
-  "Diagnostic)": "Diagnostic)",
+  "Diagnostic": "Diagnostic",
   "DIC": "DIC",
   "Disconnect": "Disconnect",
   "Disk size": "Disk size",
