+++ src/views/virtualmachines/details/tabs/configuration/disk/disk-list-page.scss
