+++ src/views/virtualmachines/details/tabs/configuration/disk/modal/DeleteDiskModal.tsx
