+++ src/views/virtualmachines/details/tabs/configuration/disk/modal/hooks/utils/utils.ts
