+++ src/views/virtualmachines/details/tabs/configuration/disk/tables/disk/HotplugLabel.tsx
