+++ src/views/virtualmachines/details/tabs/configuration/disk/tables/disk/utils/helpers.ts
