+++ src/views/virtualmachines/details/tabs/configuration/disk/tables/tables.scss
