+++ src/views/virtualmachines/details/tabs/configuration/environment/VirtualMachineEnvironmentPage.tsx
