+++ src/views/virtualmachines/details/tabs/configuration/network/NetworkInterfaceListPage.tsx
