+++ src/views/virtualmachines/details/tabs/configuration/network/copmonents/list/NetworkInterfaceList.tsx
