+++ src/views/virtualmachines/details/tabs/configuration/network/copmonents/list/NetworkInterfaceRow.tsx
