+++ src/views/virtualmachines/details/tabs/configuration/network/copmonents/modal/EditNetworkInterfaceModal.tsx
@@ -1,55 +1,46 @@
 import * as React from 'react';
 
 import VirtualMachineModel from '@kubevirt-ui/kubevirt-api/console/models/VirtualMachineModel';
-import {
-  V1Interface,
-  V1Network,
-  V1VirtualMachine,
-  V1VirtualMachineInstance,
-} from '@kubevirt-ui/kubevirt-api/kubevirt';
-import { ModalPendingChangesAlert } from '@kubevirt-utils/components/PendingChanges/ModalPendingChangesAlert/ModalPendingChangesAlert';
-import { getChangedNics } from '@kubevirt-utils/components/PendingChanges/utils/helpers';
+import { V1Interface, V1Network, V1VirtualMachine } from '@kubevirt-ui/kubevirt-api/kubevirt';
 import TabModal from '@kubevirt-utils/components/TabModal/TabModal';
+import { useKubevirtTranslation } from '@kubevirt-utils/hooks/useKubevirtTranslation';
 import { getInterfaces, getNetworks } from '@kubevirt-utils/resources/vm';
+import { NetworkPresentation } from '@kubevirt-utils/resources/vm/utils/network/constants';
+import { getNetworkInterfaceType } from '@kubevirt-utils/resources/vm/utils/network/selectors';
 import { k8sUpdate } from '@openshift-console/dynamic-plugin-sdk';
 import { Form } from '@patternfly/react-core';
-
-import NameFormField from '../../../form/NameFormField';
+import NameFormField from '@virtualmachines/details/tabs/form/NameFormField';
 
 import NetworkInterfaceMACAddressInput from './NetworkInterfaceFormFields/NetworkInterfaceMacAddressInput';
 import NetworkInterfaceModelSelect from './NetworkInterfaceFormFields/NetworkInterfaceModelSelect';
 import NetworkInterfaceNetworkSelect from './NetworkInterfaceFormFields/NetworkInterfaceNetworkSelect';
 import NetworkInterfaceTypeSelect from './NetworkInterfaceFormFields/NetworkInterfaceTypeSelect';
-import { interfaceModelType, interfaceTypeTypes } from './utils/constants';
-import {
-  generateNicName,
-  networkNameStartWithPod,
-  podNetworkExists,
-  updateVMNetworkInterface,
-} from './utils/helpers';
+import { interfaceTypeTypes } from './utils/constants';
+import { getNetworkName, networkNameStartWithPod, updateVMNetworkInterface } from './utils/helpers';
 
-type NetworkInterfaceModalProps = {
+type EditNetworkInterfaceModalProps = {
   vm: V1VirtualMachine;
   isOpen: boolean;
   onClose: () => void;
-  headerText: string;
-  vmi?: V1VirtualMachineInstance;
+  nicPresentation: NetworkPresentation;
 };
 
-const NetworkInterfaceModal: React.FC<NetworkInterfaceModalProps> = ({
+const EditNetworkInterfaceModal: React.FC<EditNetworkInterfaceModalProps> = ({
   vm,
   isOpen,
   onClose,
-  headerText,
-  vmi,
+  nicPresentation,
 }) => {
-  const [nicName, setNicName] = React.useState(generateNicName());
-  const [interfaceModel, setInterfaceModel] = React.useState(interfaceModelType.VIRTIO);
-  const [networkName, setNetworkName] = React.useState(null);
+  const { t } = useKubevirtTranslation();
+  const { network, iface } = nicPresentation;
+  const [nicName, setNicName] = React.useState(network.name);
+  const [interfaceModel, setInterfaceModel] = React.useState(iface.model);
+  const [networkName, setNetworkName] = React.useState(getNetworkName(network, t));
   const [interfaceType, setInterfaceType] = React.useState(
-    podNetworkExists(vm) ? interfaceTypeTypes.BRIDGE : interfaceTypeTypes.MASQUERADE,
+    interfaceTypeTypes[getNetworkInterfaceType(iface).toUpperCase()],
   );
-  const [interfaceMACAddress, setInterfaceMACAddress] = React.useState(undefined);
+
+  const [interfaceMACAddress, setInterfaceMACAddress] = React.useState(iface.macAddress);
 
   const [submitDisabled, setSubmitDisabled] = React.useState(true);
 
@@ -63,7 +54,10 @@ const NetworkInterfaceModal: React.FC<NetworkInterfaceModalProps> = ({
       resultNetwork.pod = {};
     }
 
-    const updatedNetworks: V1Network[] = [...(getNetworks(vm) || []), resultNetwork];
+    const updatedNetworks: V1Network[] = [
+      ...(getNetworks(vm)?.filter(({ name }) => name !== network.name) || []),
+      resultNetwork,
+    ];
 
     const resultInterface: V1Interface = {
       name: nicName,
@@ -79,10 +73,13 @@ const NetworkInterfaceModal: React.FC<NetworkInterfaceModalProps> = ({
       resultInterface.sriov = {};
     }
 
-    const updatedInterfaces: V1Interface[] = [...(getInterfaces(vm) || []), resultInterface];
+    const updatedInterfaces: V1Interface[] = [
+      ...(getInterfaces(vm)?.filter(({ name }) => name !== network.name) || []),
+      resultInterface,
+    ];
 
     return updateVMNetworkInterface(vm, updatedNetworks, updatedInterfaces);
-  }, [interfaceMACAddress, interfaceModel, interfaceType, networkName, nicName, vm]);
+  }, [interfaceMACAddress, interfaceModel, interfaceType, network.name, networkName, nicName, vm]);
 
   return (
     <TabModal
@@ -97,11 +94,10 @@ const NetworkInterfaceModal: React.FC<NetworkInterfaceModalProps> = ({
       }
       isOpen={isOpen}
       onClose={onClose}
-      headerText={headerText}
+      headerText={t('Edit network interface')}
       isDisabled={submitDisabled}
     >
       <Form>
-        {vmi && <ModalPendingChangesAlert isChanged={getChangedNics(vm, vmi)?.length > 0} />}
         <NameFormField objName={nicName} setObjName={setNicName} />
         <NetworkInterfaceModelSelect
           interfaceModel={interfaceModel}
@@ -113,6 +109,8 @@ const NetworkInterfaceModal: React.FC<NetworkInterfaceModalProps> = ({
           setNetworkName={setNetworkName}
           setInterfaceType={setInterfaceType}
           setSubmitDisabled={setSubmitDisabled}
+          isEditing
+          editInitValueNetworkName={getNetworkName(network, t)}
         />
         <NetworkInterfaceTypeSelect
           interfaceType={interfaceType}
@@ -130,4 +128,4 @@ const NetworkInterfaceModal: React.FC<NetworkInterfaceModalProps> = ({
   );
 };
 
-export default NetworkInterfaceModal;
+export default EditNetworkInterfaceModal;
