+++ src/views/virtualmachines/details/tabs/configuration/network/copmonents/modal/NetworkInterfaceFormFields/NetworkInterfaceMacAddressInput.tsx
