+++ src/views/virtualmachines/details/tabs/configuration/network/copmonents/modal/NetworkInterfaceFormFields/NetworkInterfaceTypeSelect.tsx
