+++ src/views/virtualmachines/details/tabs/configuration/network/copmonents/modal/NetworkInterfaceModal.tsx
@@ -1,47 +1,54 @@
 import * as React from 'react';
 
 import VirtualMachineModel from '@kubevirt-ui/kubevirt-api/console/models/VirtualMachineModel';
-import { V1Interface, V1Network, V1VirtualMachine } from '@kubevirt-ui/kubevirt-api/kubevirt';
+import {
+  V1Interface,
+  V1Network,
+  V1VirtualMachine,
+  V1VirtualMachineInstance,
+} from '@kubevirt-ui/kubevirt-api/kubevirt';
+import { ModalPendingChangesAlert } from '@kubevirt-utils/components/PendingChanges/ModalPendingChangesAlert/ModalPendingChangesAlert';
+import { getChangedNics } from '@kubevirt-utils/components/PendingChanges/utils/helpers';
 import TabModal from '@kubevirt-utils/components/TabModal/TabModal';
-import { useKubevirtTranslation } from '@kubevirt-utils/hooks/useKubevirtTranslation';
 import { getInterfaces, getNetworks } from '@kubevirt-utils/resources/vm';
-import { NetworkPresentation } from '@kubevirt-utils/resources/vm/utils/network/constants';
-import { getNetworkInterfaceType } from '@kubevirt-utils/resources/vm/utils/network/selectors';
 import { k8sUpdate } from '@openshift-console/dynamic-plugin-sdk';
 import { Form } from '@patternfly/react-core';
-
-import NameFormField from '../../../form/NameFormField';
+import NameFormField from '@virtualmachines/details/tabs/form/NameFormField';
 
 import NetworkInterfaceMACAddressInput from './NetworkInterfaceFormFields/NetworkInterfaceMacAddressInput';
 import NetworkInterfaceModelSelect from './NetworkInterfaceFormFields/NetworkInterfaceModelSelect';
 import NetworkInterfaceNetworkSelect from './NetworkInterfaceFormFields/NetworkInterfaceNetworkSelect';
 import NetworkInterfaceTypeSelect from './NetworkInterfaceFormFields/NetworkInterfaceTypeSelect';
-import { interfaceTypeTypes } from './utils/constants';
-import { getNetworkName, networkNameStartWithPod, updateVMNetworkInterface } from './utils/helpers';
+import { interfaceModelType, interfaceTypeTypes } from './utils/constants';
+import {
+  generateNicName,
+  networkNameStartWithPod,
+  podNetworkExists,
+  updateVMNetworkInterface,
+} from './utils/helpers';
 
-type EditNetworkInterfaceModalProps = {
+type NetworkInterfaceModalProps = {
   vm: V1VirtualMachine;
   isOpen: boolean;
   onClose: () => void;
-  nicPresentation: NetworkPresentation;
+  headerText: string;
+  vmi?: V1VirtualMachineInstance;
 };
 
-const EditNetworkInterfaceModal: React.FC<EditNetworkInterfaceModalProps> = ({
+const NetworkInterfaceModal: React.FC<NetworkInterfaceModalProps> = ({
   vm,
   isOpen,
   onClose,
-  nicPresentation,
+  headerText,
+  vmi,
 }) => {
-  const { t } = useKubevirtTranslation();
-  const { network, iface } = nicPresentation;
-  const [nicName, setNicName] = React.useState(network.name);
-  const [interfaceModel, setInterfaceModel] = React.useState(iface.model);
-  const [networkName, setNetworkName] = React.useState(getNetworkName(network, t));
+  const [nicName, setNicName] = React.useState(generateNicName());
+  const [interfaceModel, setInterfaceModel] = React.useState(interfaceModelType.VIRTIO);
+  const [networkName, setNetworkName] = React.useState(null);
   const [interfaceType, setInterfaceType] = React.useState(
-    interfaceTypeTypes[getNetworkInterfaceType(iface).toUpperCase()],
+    podNetworkExists(vm) ? interfaceTypeTypes.BRIDGE : interfaceTypeTypes.MASQUERADE,
   );
-
-  const [interfaceMACAddress, setInterfaceMACAddress] = React.useState(iface.macAddress);
+  const [interfaceMACAddress, setInterfaceMACAddress] = React.useState(undefined);
 
   const [submitDisabled, setSubmitDisabled] = React.useState(true);
 
@@ -55,10 +62,7 @@ const EditNetworkInterfaceModal: React.FC<EditNetworkInterfaceModalProps> = ({
       resultNetwork.pod = {};
     }
 
-    const updatedNetworks: V1Network[] = [
-      ...(getNetworks(vm)?.filter(({ name }) => name !== network.name) || []),
-      resultNetwork,
-    ];
+    const updatedNetworks: V1Network[] = [...(getNetworks(vm) || []), resultNetwork];
 
     const resultInterface: V1Interface = {
       name: nicName,
@@ -74,13 +78,10 @@ const EditNetworkInterfaceModal: React.FC<EditNetworkInterfaceModalProps> = ({
       resultInterface.sriov = {};
     }
 
-    const updatedInterfaces: V1Interface[] = [
-      ...(getInterfaces(vm)?.filter(({ name }) => name !== network.name) || []),
-      resultInterface,
-    ];
+    const updatedInterfaces: V1Interface[] = [...(getInterfaces(vm) || []), resultInterface];
 
     return updateVMNetworkInterface(vm, updatedNetworks, updatedInterfaces);
-  }, [interfaceMACAddress, interfaceModel, interfaceType, network.name, networkName, nicName, vm]);
+  }, [interfaceMACAddress, interfaceModel, interfaceType, networkName, nicName, vm]);
 
   return (
     <TabModal
@@ -95,10 +96,11 @@ const EditNetworkInterfaceModal: React.FC<EditNetworkInterfaceModalProps> = ({
       }
       isOpen={isOpen}
       onClose={onClose}
-      headerText={t('Edit network interface')}
+      headerText={headerText}
       isDisabled={submitDisabled}
     >
       <Form>
+        {vmi && <ModalPendingChangesAlert isChanged={getChangedNics(vm, vmi)?.length > 0} />}
         <NameFormField objName={nicName} setObjName={setNicName} />
         <NetworkInterfaceModelSelect
           interfaceModel={interfaceModel}
@@ -110,8 +112,6 @@ const EditNetworkInterfaceModal: React.FC<EditNetworkInterfaceModalProps> = ({
           setNetworkName={setNetworkName}
           setInterfaceType={setInterfaceType}
           setSubmitDisabled={setSubmitDisabled}
-          isEditing
-          editInitValueNetworkName={getNetworkName(network, t)}
         />
         <NetworkInterfaceTypeSelect
           interfaceType={interfaceType}
@@ -129,4 +129,4 @@ const EditNetworkInterfaceModal: React.FC<EditNetworkInterfaceModalProps> = ({
   );
 };
 
-export default EditNetworkInterfaceModal;
+export default NetworkInterfaceModal;
