+++ src/views/virtualmachines/details/tabs/configuration/network/copmonents/modal/utils/constants.ts
