+++ src/views/virtualmachines/details/tabs/configuration/network/copmonents/modal/utils/mac-validation.ts
