+++ src/views/virtualmachines/details/tabs/configuration/network/hooks/useNetworkColumns.ts
