+++ src/views/virtualmachines/details/tabs/configuration/network/network-interface-list-page.scss
