+++ src/views/virtualmachines/details/tabs/configuration/scheduling/VirtualMachineSchedulingPage.tsx
