+++ src/views/virtualmachines/details/tabs/configuration/scheduling/components/Affinity.tsx
