+++ src/views/virtualmachines/details/tabs/configuration/scheduling/components/DedicatedResources.tsx
