+++ src/views/virtualmachines/details/tabs/configuration/scheduling/components/EvictionStrategy.tsx
