+++ src/views/virtualmachines/details/tabs/configuration/scheduling/components/NodeSelector.tsx
