+++ src/views/virtualmachines/details/tabs/configuration/scheduling/components/SchedulingSection/SchedulingSection.scss
