+++ src/views/virtualmachines/details/tabs/configuration/scripts/scripts-tab.scss
